# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics ERP API — Sales & CRM, Project Management, Inventory & Warehouse Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List and filter leads
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: owner
          in: query
          schema:
            type: string
        - name: territory
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /leads/{id}/assign:
    post:
      operationId: assignLead
      tags: [leads]
      summary: Assign lead and compute score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadAssignInput"
      responses:
        "200":
          description: Lead assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── AUTH ─────────────────────────────────────────────────────────────────
...
    Lead:
      type: object
      required: [id, source, status, createdAt]
      properties:
        id:
          type: integer
        source:
          type: string
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        territory:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        productInterest:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        score:
          type: ["integer", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadInput:
      type: object
      required: [source, status]
      properties:
        source:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
        territory:
          type: string
        productInterest:
          type: string
        estimatedValue:
          type: number
        status:
          type: string
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
        territory:
          type: string
        productInterest:
          type: string
        estimatedValue:
          type: number
        status:
          type: string
        notes:
          type: string
        ownerId:
          type: integer
    LeadAssignInput:
      type: object
      required: [ownerId]
      properties:
        ownerId:
          type: integer

    # ── QUOTATIONS ────────────────────────────────────────────────────────────
...